replacing float by double for temperature

+++ prediction-engine/src/main/java/fr/lernejo/prediction/Temperature.java
@@ -1,4 +1,4 @@
 package fr.lernejo.prediction;
 
-public record Temperature(String country, String date, float temperature) {
+public record Temperature(String country, String date, double temperature) {
 }
